Move utils/router.ts to common/router.ts
